% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{bake}
\alias{bake}
\title{Generate and stores data from a biobrick in your \code{bblib()}}
\usage{
bake(brick)
}
\arguments{
\item{brick}{the brick you would like to generate}
}
\description{
Generate and stores data from a biobrick in your \code{bblib()}
}
