% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbassets.R
\name{bbassets}
\alias{bbassets}
\title{Return a list of tables for a brick}
\usage{
bbassets(brick)
}
\arguments{
\item{brick}{the name of the brick to load}
}
\value{
A named list where each element is a character string representing
the full path to a specific asset file. The keys of the list correspond to
the types of asset files available.
}
\description{
Return a list of tables for a brick
}
