% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{bbinstall}
\alias{bbinstall}
\title{Install a brickref}
\usage{
bbinstall(brickref)
}
\arguments{
\item{brickref}{either a repo name from github.com/biobricks-ai or a full github url.}
}
\description{
Install a brickref
}
