% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_load_arrow}
\alias{brick_load_arrow}
\title{A helper to use brick_load with the arrow package}
\usage{
brick_load_arrow(brick)
}
\arguments{
\item{brick}{the name of the brick to load}
}
\description{
A helper to use brick_load with the arrow package
}
