% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_load_sqlite}
\alias{brick_load_sqlite}
\title{A helper to use brick_load with a custom sqlite loader. opens a sqlite connection
and closes it when parent exits.}
\usage{
brick_load_sqlite(brick, env = parent.frame())
}
\arguments{
\item{brick}{the name of the brick to load}

\item{env}{when @param env exits, biobricks closes the sqlite connection}
}
\description{
A helper to use brick_load with a custom sqlite loader. opens a sqlite connection
and closes it when parent exits.
}
