% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_ls}
\alias{brick_ls}
\title{Get the files in a brick}
\usage{
brick_ls(brick, .p = NULL, rel)
}
\arguments{
\item{brick}{the brick to get files for}

\item{.p}{a predicate to filter returned files (must return true on file path)}

\item{rel}{a predicate to filter returned files (must return true on file path)}
}
\description{
Get the files in a brick
}
