% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{brick_remove}
\alias{brick_remove}
\title{removes biobrick from bblib
removes biobrick org if no other bricks under org
removes git submodule}
\usage{
brick_remove(brick)
}
\arguments{
\item{brick}{the brick to remove}
}
\description{
removes biobrick from bblib
removes biobrick org if no other bricks under org
removes git submodule
}
