% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick_repro.R
\name{brick_repro}
\alias{brick_repro}
\title{Generate and stores data from a biobrick in your \code{bblib()}
This is a wrapper for dvc repro.}
\usage{
brick_repro(brick, env = "docker", image = "insilica/biobricks:latest")
}
\arguments{
\item{brick}{the brick you would like to generate}

\item{env}{"docker" for using a docker image & "system" for your local machine (must have dvc)}

\item{image}{if "docker" env which docker image to use}
}
\description{
Generate and stores data from a biobrick in your \code{bblib()}
This is a wrapper for dvc repro.
}
\examples{
\dontrun{
biobricks::install("https://github.com/biobricks-ai/clinvar.git")
biobricks::brick_repro("clinvar")
}
}
