% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{brick_update}
\alias{brick_update}
\title{Update a brick in 3 steps:
\enumerate{
\item stashes all local changes
\item installs the most recent from origin
\item drops the stash
}}
\usage{
brick_update(brick)
}
\arguments{
\item{brick}{the brick to update}
}
\description{
Update a brick in 3 steps:
\enumerate{
\item stashes all local changes
\item installs the most recent from origin
\item drops the stash
}
}
