% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{brickpull}
\alias{brickpull}
\title{pull data for a brick from biobricks.ai}
\usage{
brickpull(brick)
}
\arguments{
\item{brick}{pull data from this brick}
}
\description{
pull data for a brick from biobricks.ai
}
\examples{
\dontrun{
biobricks::brickpull("clinvar")
}
}
