% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_run}
\alias{docker_run}
\title{Run a cmd in the docker container}
\usage{
docker_run(cmd, wd = NA)
}
\arguments{
\item{cmd}{The command to run}

\item{wd}{Working directory context to execute cmd}
}
\description{
Run a cmd in the docker container
}
