% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fs_list}
\alias{fs_list}
\title{creates a nested list of values by path # TODO replace this with something simpler}
\usage{
fs_list(x, path)
}
\arguments{
\item{x}{a set of values to make into nested list}

\item{path}{filesytem paths to nest}
}
\description{
creates a nested list of values by path # TODO replace this with something simpler
}
