% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fs_recurse}
\alias{fs_recurse}
\title{creates a nested list of values by path}
\usage{
fs_recurse(x, path = names(x))
}
\arguments{
\item{x}{a set of values to make into nested list}

\item{path}{filesytem paths to nest}
}
\description{
creates a nested list of values by path
}
