% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.biobricks}
\alias{install.biobricks}
\title{Installs a biobricks-ai maintained brick}
\usage{
install.biobricks(brick)
}
\arguments{
\item{brick}{a brick from biobricks-ai, see https://github.com/biobricks-ai/{brick}}
}
\description{
Installs a biobricks-ai maintained brick
}
