% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Installs a brick from a github repo}
\usage{
install(url)
}
\arguments{
\item{url}{a url like https://github.com/biobricks-ai/clinvar.git}
}
\description{
Installs a brick from a github repo
}
