% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gh}
\alias{install_gh}
\title{Installs a brick from a github repo}
\usage{
install_gh(repo)
}
\arguments{
\item{repo}{string with owner/repo eg. "biobricks-ai/clinvar"}
}
\description{
Installs a brick from a github repo
}
