% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lazy}
\alias{lazy}
\title{attempts to parse a data directory into dplyr lazy tables}
\usage{
lazy(files)
}
\arguments{
\item{files}{an array of files to parse}
}
\description{
attempts to parse a data directory into dplyr lazy tables
}
