% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{local_bblib}
\alias{local_bblib}
\title{creates a biobricks library in a temporary directory. Useful for development}
\usage{
local_bblib(env = parent.frame())
}
\arguments{
\item{env}{environment in which to create a temporary biobicks library}
}
\description{
creates a biobricks library in a temporary directory. Useful for development
}
