% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ptree}
\alias{ptree}
\title{creates a nested list of values by path}
\usage{
ptree(
  x,
  p = names(x),
  pd = fs::path_dir(p),
  pr = purrr::map2(p, pd, fs::path_rel)
)
}
\arguments{
\item{x}{a set of values to make into nested list}

\item{p}{filesytem paths to nest and use for names}

\item{pd}{'path_dir' used for recursion}

\item{pr}{'path_rel' used for recursion}
}
\description{
creates a nested list of values by path
}
