% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{resolve}
\alias{resolve}
\title{Finds all the bricks in \code{bblib()} matching @param brick}
\usage{
resolve(brick, rel = "/")
}
\arguments{
\item{brick}{the brick you want to find}

\item{rel}{resolve relative to what path}
}
\description{
Finds all the bricks in \code{bblib()} matching @param brick
}
