% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{tbls}
\alias{tbls}
\title{Return a list of tables for a brick
TODO improve loading. How do we handle sqlite? other loading functions?}
\usage{
tbls(brick, load = arrow::open_dataset)
}
\arguments{
\item{brick}{the name of the brick to import}

\item{load}{how should files be loaded in R?}
}
\description{
Return a list of tables for a brick
TODO improve loading. How do we handle sqlite? other loading functions?
}
