% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{update.biobricks}
\alias{update.biobricks}
\title{Update a brick
#TODO #4 add a test for update.biobricks, make force actually work}
\usage{
\method{update}{biobricks}(brick, force = F)
}
\arguments{
\item{brick}{the brick to update}

\item{force}{throw out local changes}
}
\description{
Update a brick
#TODO #4 add a test for update.biobricks, make force actually work
}
