% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbassets.R
\name{bbassets}
\alias{bbassets}
\title{Return a list of tables for a brick}
\usage{
bbassets(brick)
}
\arguments{
\item{brick}{the name of the brick to load}
}
\description{
Return a list of tables for a brick
}
