% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{brick_install}
\alias{brick_install}
\title{Installs a biobricks-ai maintained brick}
\usage{
brick_install(brick)
}
\arguments{
\item{brick}{a brick from biobricks-ai, see https://github.com/biobricks-ai/{brick}}
}
\description{
Installs a biobricks-ai maintained brick
}
