% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_load}
\alias{brick_load}
\title{Return a list of tables for a brick}
\usage{
brick_load(brick, .p, .l)
}
\arguments{
\item{brick}{the name of the brick to load}

\item{.p}{which files should be loaded this way?}

\item{.l}{which loading function should be used?}
}
\description{
Return a list of tables for a brick
}
