% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_load_sqlite}
\alias{brick_load_sqlite}
\title{A helper to use brick_load with a custom sqlite loader. opens a sqlite connection
and closes it when parent exits.}
\usage{
brick_load_sqlite(
  brick,
  .p = partial(grepl, pat = "(\\\\.db$|\\\\.sqlite)", i = T),
  env = parent.frame()
)
}
\arguments{
\item{brick}{the name of the brick to load}

\item{.p}{default to grepl on sqlite names like .db and .sqlite}

\item{env}{when @param env exits, biobricks closes the sqlite connection}
}
\description{
A helper to use brick_load with a custom sqlite loader. opens a sqlite connection
and closes it when parent exits.
}
