% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{brick_ls}
\alias{brick_ls}
\title{Get the files in a brick}
\usage{
brick_ls(brick, regex)
}
\arguments{
\item{brick}{the brick to get files for}

\item{regex}{the brick to get files for}
}
\description{
Get the files in a brick
}
