% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{brick_path}
\alias{brick_path}
\title{Finds all the bricks in \code{bblib()} matching @param brick}
\usage{
brick_path(brick, path = "")
}
\arguments{
\item{brick}{the brick you want to find}

\item{path}{retrieve path relative to the brick directory}
}
\description{
Finds all the bricks in \code{bblib()} matching @param brick
}
\examples{
\dontrun{
biobricks::resolve("clinvar")
}
}
