% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Installs a brick from a github repo}
\usage{
install(url, repo)
}
\arguments{
\item{url}{a url like https://github.com/biobricks-ai/clinvar.git}

\item{repo}{string with owner/repo eg. "biobricks-ai/clinvar"}
}
\description{
Installs a brick from a github repo
}
