% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gh}
\alias{install_gh}
\title{Installs a brick from a github repo}
\usage{
install_gh(owner, repo)
}
\arguments{
\item{owner}{a github owner "biobricks-ai" in https://github.com/biobricks-ai/clinvar}

\item{repo}{a github repo "clinvar" in https://github.com/biobricks-ai/clinvar}
}
\description{
Installs a brick from a github repo
}
