% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{update_brick}
\alias{update_brick}
\title{Update a brick
#TODO #4 add a test for update_brick}
\usage{
update_brick(brick)
}
\arguments{
\item{brick}{the brick to update}
}
\description{
Update a brick
#TODO #4 add a test for update_brick
}
